requirements.txt
cosine_similarity
language_check
textstat
scipy
numpy
nltk
jiwer
tqdm
pandas
bs4
dask
matplotlib
tikzplotlib
sklearn
numba
gensim
textblob
spacy
scrapy
pymongo
torch
transformers
h5py
boto3
tensorboard
hydra-core
sentencepiece
git+https://github.com/huggingface/transformers.git
peft
fire
bitsandbytes
accelerate
